Add constructors to SoldProduct entity
Implements SoldProduct constructors and minor changes on its repository.

+++ cccpharma/src/main/java/br/edu/ufcg/cccpharma/soldProduct/SoldProductRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 @Repository
 public interface SoldProductRepository extends JpaRepository<SoldProduct, Long> {
 
-	SoldProduct findById(int Id);
+	SoldProduct findById(long Id);
 
 }
